refactor: cleanup, extract regroupImports
